+++ ikea-zigbee-drivers/Ikea_E2204.groovy
@@ -7,7 +7,7 @@ import groovy.transform.CompileStatic
 import groovy.transform.Field
 
 @Field static final String DRIVER_NAME = 'IKEA Tretakt Smart Plug (E2204)'
-@Field static final String DRIVER_VERSION = '4.0.0'
+@Field static final String DRIVER_VERSION = '4.1.0'
 
 // Fields for capability.HealthCheck
 import groovy.time.TimeCategory
@@ -51,7 +51,7 @@ metadata {
             name: 'helpInfo', type: 'hidden',
             title: '''
             <div style="min-height:55px; background:transparent url('https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/img/Ikea_E2204.webp') no-repeat left center;background-size:auto 55px;padding-left:60px">
-                IKEA Tretakt Smart Plug (E2204) <small>v4.0.0</small><br>
+                IKEA Tretakt Smart Plug (E2204) <small>v4.1.0</small><br>
                 <small><div>
                 • <a href="https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/#tretakt-smart-plug-e2204" target="_blank">device details</a><br>
                 • <a href="https://community.hubitat.com/t/release-ikea-zigbee-drivers/123853" target="_blank">community page</a><br>
@@ -93,6 +93,12 @@ metadata {
             name: 'childLock', type: 'bool',
             title: 'Child lock',
             description: '<small>Lock physical button, safeguarding against accidental operation.</small>',
+            defaultValue: false
+        )
+        input(
+            name: 'darkMode', type: 'bool',
+            title: 'Dark mode',
+            description: '<small>Turn off LED indicators on the device, ensuring total darkness.</small>',
             defaultValue: false
         )
         
@@ -147,6 +153,13 @@ List<String> updated(boolean auto = false) {
     }
     log_info "🛠️ childLock = ${childLock}"
     cmds += zigbee.writeAttribute(0xFC85, 0x0000, 0x10, childLock ? 0x01 : 0x00, [mfgCode:'0x117C'])
+    
+    if (darkMode == null) {
+        darkMode = false
+        device.updateSetting 'darkMode', [value:darkMode, type:'bool']
+    }
+    log_info "🛠️ darkMode = ${darkMode}"
+    cmds += zigbee.writeAttribute(0xFC85, 0x0001, 0x10, darkMode ? 0x01 : 0x00, [mfgCode:'0x117C'])
     
     // Preferences for capability.HealthCheck
     schedule HEALTH_CHECK.schedule, 'healthCheck'
@@ -226,7 +239,9 @@ void configure(boolean auto = false) {
     cmds += zigbee.readAttribute(0x0000, 0x0007)  // PowerSource
 
     // Query Basic cluster attributes
-    cmds += zigbee.readAttribute(0x0000, [0x0001, 0x0003, 0x0004, 0x0005, 0x000A, 0x4000]) // ApplicationVersion, HWVersion, ManufacturerName, ModelIdentifier, ProductCode, SWBuildID
+    cmds += zigbee.readAttribute(0x0000, [0x0001, 0x0003, 0x0004, 0x4000]) // ApplicationVersion, HWVersion, ManufacturerName, SWBuildID
+    cmds += zigbee.readAttribute(0x0000, [0x0005]) // ModelIdentifier
+    cmds += zigbee.readAttribute(0x0000, [0x000A]) // ProductCode
     utils_sendZigbeeCommands cmds
 
     log_info 'Configuration done; refreshing device current state in 7 seconds ...'
@@ -252,6 +267,7 @@ void refresh(boolean auto = false) {
     
     // Refresh for devices.Ikea_E2204
     cmds += zigbee.readAttribute(0xFC85, 0x0000, [mfgCode:'0x117C'] ) // ChildLock
+    cmds += zigbee.readAttribute(0xFC85, 0x0001, [mfgCode:'0x117C'] ) // DarkMode
     
     // Refresh for capability.ZigbeeGroups
     cmds += "he raw 0x${device.deviceNetworkId} 0x01 0x${device.endpointId} 0x0004 {0143 02 00}"  // Get groups membership
@@ -374,7 +390,7 @@ void parse(String description) {
                 case 0x00: newValue = 'TURN_POWER_OFF'; break
                 case 0x01: newValue = 'TURN_POWER_ON'; break
                 case 0xFF: newValue = 'RESTORE_PREVIOUS_STATE'; break
-                default: log_warn "Received attribute value: powerOnBehavior=${msg.value}"; return
+                default: log_warn "Received unexpected attribute value: PowerOnBehavior=${msg.value}"; return
             }
             powerOnBehavior = newValue
             device.updateSetting 'powerOnBehavior', [value:newValue, type:'enum']
@@ -383,7 +399,7 @@ void parse(String description) {
         
         // Other events that we expect but are not usefull for capability.Switch behavior
         case { contains it, [clusterInt:0x0006, commandInt:0x07] }:
-            utils_processedZclMessage 'Configure Reporting Response', "attribute=switch, data=${msg.data}"
+            utils_processedZclMessage 'Configure Reporting Response', "attribute=OnOff, data=${msg.data}"
             return
         case { contains it, [clusterInt:0x0006, commandInt:0x04] }: // Write Attribute Response
         case { contains it, [clusterInt:0x0006, commandInt:0x06, isClusterSpecific:false, direction:'01'] }: // Configure Reporting Command
@@ -399,9 +415,15 @@ void parse(String description) {
             utils_processedZclMessage 'Read Attributes Response', "ChildLock=${msg.value}"
             return
         
+        // Read Attributes: DarkMode
+        case { contains it, [clusterInt:0xFC85, commandInt:0x01, attrInt:0x0001] }:
+            darkMode = msg.value == '01'
+            device.updateSetting 'darkMode', [darkMode:childLock, type:'bool']
+            utils_processedZclMessage 'Read Attributes Response', "DarkMode=${msg.value}"
+            return
+        
         // Write Attributes Response
         case { contains it, [endpointInt:0x01, clusterInt:0xFC85, commandInt:0x04, isClusterSpecific:false, isManufacturerSpecific:true, manufacturerId:'117C'] }:
-            log_info 'Child lock successfully configured!'
             return
         
         // Events for capability.HealthCheck