+++ ikea-zigbee-drivers/Philips_RWL022.groovy
@@ -7,7 +7,7 @@ import groovy.transform.CompileStatic
 import groovy.transform.Field
 
 @Field static final String DRIVER_NAME = 'Philips Hue Dimmer Switch (RWL022)'
-@Field static final String DRIVER_VERSION = '4.0.0'
+@Field static final String DRIVER_VERSION = '4.1.0'
 
 // Fields for capability.HealthCheck
 import groovy.time.TimeCategory
@@ -56,7 +56,7 @@ metadata {
             name: 'helpInfo', type: 'hidden',
             title: '''
             <div style="min-height:55px; background:transparent url('https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/img/Philips_RWL022.webp') no-repeat left center;background-size:auto 55px;padding-left:60px">
-                Philips Hue Dimmer Switch (RWL022) <small>v4.0.0</small><br>
+                Philips Hue Dimmer Switch (RWL022) <small>v4.1.0</small><br>
                 <small><div>
                 • <a href="https://dan-danache.github.io/hubitat/ikea-zigbee-drivers/#philips-hue-dimmer-switch-rwl022" target="_blank">device details</a><br>
                 • <a href="https://community.hubitat.com/t/release-ikea-zigbee-drivers/123853" target="_blank">community page</a><br>
@@ -226,7 +226,9 @@ void configure(boolean auto = false) {
     sendEvent name:'numberOfButtons', value:numberOfButtons, descriptionText:"Number of buttons is ${numberOfButtons}"
 
     // Query Basic cluster attributes
-    cmds += zigbee.readAttribute(0x0000, [0x0001, 0x0003, 0x0004, 0x0005, 0x000A, 0x4000]) // ApplicationVersion, HWVersion, ManufacturerName, ModelIdentifier, ProductCode, SWBuildID
+    cmds += zigbee.readAttribute(0x0000, [0x0001, 0x0003, 0x0004, 0x4000]) // ApplicationVersion, HWVersion, ManufacturerName, SWBuildID
+    cmds += zigbee.readAttribute(0x0000, [0x0005]) // ModelIdentifier
+    cmds += zigbee.readAttribute(0x0000, [0x000A]) // ProductCode
     utils_sendZigbeeCommands cmds
 
     log_info 'Configuration done; refreshing device current state in 7 seconds ...'